+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Versions
 
 ## Unreleased
+- Exclude assemblies that set DefinedTypes to null instead of an empty array
+  to prevent failures on GetTypes(). This issue occurred with the
+  Microsoft.WindowsAzure.Storage.dll assembly.
 
 ## 4.4.0.0